Enum odbc_safe::Return

#[must_use]

pub enum Return<T, E = ()> {
    Success(T),
    Info(T),
    Error(E),
}

Holds result and indicates the overall success or failure of a function.

Variants
Success(T)

The function has been executed successfully. Holds result.

Info(T)

The function has been executed successfully. There have been warnings. Holds result.

Error(E)

An error occured.

Methods

impl<T, E> Return<T, E>

pub fn map<F, U>(self, f: F) -> Return<U, E> where
F: FnOnce(T) -> U,

Maps a Return<T,E> to Return<U,E> by applying a function to a contained Success or Info value, leaving an Error value untouched.

pub fn map_error<F, U>(self, f: F) -> Return<T, U> where
F: FnOnce(E) -> U,

Maps a Return<T,E> to Result<T,U> by applying a function to a contained Error value, leaving aSuccessor anInfo` value untouched.

pub fn unwrap(self) -> T

Unwraps the result, yielding the content of Success or Info

pub fn success<U>(self) -> Result<T, U> where
U: From<E>,

Transforms the Return<T,E> into a Result<T,U>, mapping Success(v) | Info(v) to Ok(v) and Error(err) to Err(err.into()).