Struct odbc_safe::Environment

pub struct Environment<V> { /* fields omitted */ }

An Environment is a global context, in which to access data.

Associated with an Environment is any information that is global in nature, such as:

The Environment's state

The current environment-level diagnostics

The handles of connections currently allocated on the environment

The current stetting of each environment attribute

See: Environment Handles in the ODBC Reference

Methods

impl<V> Environment<V>

pub fn as_raw(&self) -> SQLHENV

Provides access to the raw ODBC environment handle.

impl<V: Version> Environment<V>

pub fn data_sources(
&mut self,
direction: FetchOrientation,
server_name: &mut [u8],
description: &mut [u8]
) -> ReturnOption<(SQLSMALLINT, SQLSMALLINT)>

Fills buffers with information about the available datasources

A 32 / 64 Bit Application will only return information about either 32 or 64 Bit DataSources.

Returns

(server_name_length, description_length)

See SQLDataSources

pub fn drivers(
&mut self,
direction: FetchOrientation,
description: &mut [u8],
attributes: &mut [u8]
) -> ReturnOption<(SQLSMALLINT, SQLSMALLINT)>

Fills buffers with information about the available datasources

A 32 / 64 Bit Application will only return information about either 32 or 64 Bit DataSources.

Returns

(description_length, attributes_length)

See SQLDrivers

impl Environment<NoVersion>

pub fn new() -> Return<Self>

Allocates a new Environment

pub fn declare_version<V: Version>(
self
) -> Return<Environment<V>, Environment<NoVersion>>

Before an application allocates a connection which specification it follows. Currently these bindings only support ODBC 3.x.

It is valid to specify ODBC 3.x even then connecting to an ODBC 2.x driver. Applications must however avoid calling 3.x functionality on 2.x drivers. Since drivers are connected at runtime, these kind of errors can not be catched by the type system.

pub fn declare_version_3_8(
self
) -> Return<Environment<Odbc3m8>, Environment<NoVersion>>

Before an application allocates a connection which specification it follows. Currently these bindings only support ODBC 3.x.

It is valid to specify ODBC 3.x even then connecting to an ODBC 2.x driver. Applications must however avoid calling 3.x functionality on 2.x drivers. Since drivers are connected at runtime, these kind of errors can not be catched by the type system.

This method is a shorthand for declare_version::<Odbc3m8>.

pub fn declare_version_3(
self
) -> Return<Environment<Odbc3>, Environment<NoVersion>>

Before an application allocates a connection which specification it follows. Currently these bindings only support ODBC 3.x.

It is valid to specify ODBC 3.x even then connecting to an ODBC 2.x driver. Applications must however avoid calling 3.x functionality on 2.x drivers. Since drivers are connected at runtime, these kind of errors can not be catched by the type system.

This method is a shorthand for declare_version::<Odbc3>.