Type Definition odbc_safe::Connection

type Connection<'env> = DataSource<'env, Connected<'env>>;

Connection can be used as a shorthand for a DataSource in Connected state.

Methods

impl<'env> Connection<'env>

pub fn disconnect(
self
) -> Return<DataSource<'env, Unconnected<'env>>, Connection<'env>>

When an application has finished using a data source, it calls disconnect. disconnect disconnects the driver from the data source.

See [Disconnecting from a Data Source or Driver][1]

See [SQLDisconnect Function][2] [1]: https://docs.microsoft.com/sql/odbc/reference/develop-app/disconnecting-from-a-data-source-or-driver [2]: https://docs.microsoft.com/sql/odbc/reference/syntax/sqldisconnect-function

pub fn is_read_only(&mut self) -> Return<bool>

true if the data source is set to READ ONLY mode, false otherwise.