Trait odbc_safe::CDataType

pub unsafe trait CDataType {
    fn c_data_type() -> SqlCDataType;

    fn sql_ptr(&self) -> *const c_void;

    fn mut_sql_ptr(&mut self) -> SQLPOINTER;

    fn buffer_len(&self) -> SQLLEN;
}

See [C Data Types in ODBC][1] [1]: https://docs.microsoft.com/sql/odbc/reference/develop-app/c-data-types-in-odbc

Required Methods

fn c_data_type() -> SqlCDataType

C Data type of the buffer returned by mut_sql_ptr().

fn sql_ptr(&self) -> *const c_void

Const sql pointer

fn mut_sql_ptr(&mut self) -> SQLPOINTER

Pointer to the buffer in which should be filled with data.

fn buffer_len(&self) -> SQLLEN

Length of the buffer returned by mut_sql_ptr() in bytes.